+++ crates/symbolicator/src/services/download/s3.rs
@@ -176,14 +176,13 @@ impl S3Downloader {
                 // - Otherwise, a 403 ("access denied") is returned.
                 log::debug!("Skipping response from s3://{}/{}: {}", bucket, &key, err);
 
-                if let RusotoError::Unknown(response) = &err {
-                    if response.status.is_client_error() {
-                        return Ok(DownloadStatus::NotFound);
+                match &err {
+                    RusotoError::Service(_) => return Ok(DownloadStatus::NotFound),
+                    RusotoError::Unknown(response) if response.status.is_client_error() => {
+                        return Ok(DownloadStatus::NotFound)
                     }
+                    _ => return Err(err.into()),
                 }
-                // TODO: use this once we start writing DownloadErrors to cache
-                // return Err(err.into());
-                return Ok(DownloadStatus::NotFound);
             }
             Err(_) => {
                 // Timed out