+++ crates/symbolicator/src/services/objects/data_cache.rs
@@ -23,7 +23,7 @@ use tempfile::tempfile_in;
 use crate::cache::{CacheKey, CacheStatus};
 use crate::logging::LogError;
 use crate::services::cacher::{CacheItemRequest, CachePath};
-use crate::services::download::{DownloadStatus, RemoteDif};
+use crate::services::download::{DownloadError, DownloadStatus, RemoteDif};
 use crate::types::{ObjectId, Scope};
 use crate::utils::compression::decompress_object_file;
 use crate::utils::futures::BoxedFuture;
@@ -68,12 +68,15 @@ impl ObjectHandle {
     }
 
     pub fn parse(&self) -> Result<Option<Object<'_>>, ObjectError> {
-        match self.status {
+        // Interestingly all usages of parse() check to make sure that self.status == Positive before
+        // actually invoking it.
+        match &self.status {
             CacheStatus::Positive => Ok(Some(Object::parse(&self.data)?)),
             CacheStatus::Negative => Ok(None),
             CacheStatus::Malformed(_) => Err(ObjectError::Malformed),
-            // TODO: replace with ObjectError::Error once we want to start writing CacheSpecificErrors to cache
-            CacheStatus::CacheSpecificError(_) => Err(ObjectError::Malformed),
+            CacheStatus::CacheSpecificError(message) => Err(ObjectError::Download(
+                DownloadError::CachedError(message.clone()),
+            )),
         }
     }
 
@@ -138,6 +141,12 @@ impl CacheItemRequest for FetchFileDataRequest {
     ///
     /// If the object file did not exist on the source a [`CacheStatus::Negative`] will be
     /// returned.
+    ///
+    /// If there was an error downloading the object file, an `Ok` with
+    /// [`CacheStatus::CacheSpecificError`] is returned.
+    ///
+    /// If the object file did not exist on the source an `Ok` with [`CacheStatus::Negative`] will
+    /// be returned.
     fn compute(&self, path: &Path) -> BoxedFuture<Result<CacheStatus, Self::Error>> {
         let cache_key = self.get_cache_key();
         log::trace!("Fetching file data for {}", cache_key);
@@ -168,8 +177,8 @@ impl CacheItemRequest for FetchFileDataRequest {
                 }
 
                 Err(e) => {
-                    log::error!("Error while downloading file: {}", LogError(&e));
-                    return Ok(CacheStatus::Malformed(e.for_cache()));
+                    log::debug!("Error while downloading file: {}", LogError(&e));
+                    return Ok(CacheStatus::CacheSpecificError(e.for_cache()));
                 }
 
                 Ok(DownloadStatus::Completed) => {
@@ -319,7 +328,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_negative_cache_server_error() {
+    async fn test_download_error_cache_server_error() {
         test::setup();
 
         let server = test::FailingSymbolServer::new();
@@ -338,8 +347,8 @@ mod tests {
             };
 
             // for each of the different symbol sources, we assert that:
-            // * we get a negative cache result no matter how often we try
-            // * we hit the symbol source exactly once for the initial request
+            // * we get a cache-specific error no matter how often we try
+            // * we hit the symbol source exactly once for the initial request, followed by 3 retries
             // * the second try should *not* hit the symbol source, but should rather be served by the cache
 
             // server rejects the request (500)
@@ -348,10 +357,20 @@ mod tests {
                 ..find_object
             };
             let result = objects_actor.find(find_object.clone()).await.unwrap();
-            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
-            assert_eq!(server.accesses(), 1);
+            assert_eq!(
+                result.meta.clone().unwrap().status,
+                CacheStatus::CacheSpecificError(String::from(
+                    "failed to download: 500 Internal Server Error"
+                ))
+            );
+            assert_eq!(server.accesses(), 1 + 3); // 1 initial attempt + 3 retries
             let result = objects_actor.find(find_object.clone()).await.unwrap();
-            assert_eq!(result.meta.unwrap().status, CacheStatus::Negative);
+            assert_eq!(
+                result.meta.unwrap().status,
+                CacheStatus::CacheSpecificError(String::from(
+                    "failed to download: 500 Internal Server Error"
+                ))
+            );
             assert_eq!(server.accesses(), 0);
         })
         .await;
@@ -397,7 +416,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_negative_cache_timeout() {
+    async fn test_download_error_cache_timeout() {
         test::setup();
 
         let server = test::FailingSymbolServer::new();
@@ -416,8 +435,8 @@ mod tests {
             };
 
             // for each of the different symbol sources, we assert that:
-            // * we get a negative cache result no matter how often we try
-            // * we hit the symbol source exactly once for the initial request
+            // * we get a cache-specific error no matter how often we try
+            // * we hit the symbol source exactly once for the initial request, followed by 3 retries
             // * the second try should *not* hit the symbol source, but should rather be served by the cache
 
             // server accepts the request, but never sends any reply (timeout)
@@ -428,13 +447,13 @@ mod tests {
             let result = objects_actor.find(find_object.clone()).await.unwrap();
             assert_eq!(
                 result.meta.unwrap().status,
-                CacheStatus::Malformed(String::from("download was cancelled"))
+                CacheStatus::CacheSpecificError(String::from("download was cancelled"))
             );
             assert_eq!(server.accesses(), 1);
             let result = objects_actor.find(find_object.clone()).await.unwrap();
             assert_eq!(
                 result.meta.unwrap().status,
-                CacheStatus::Malformed(String::from("download was cancelled"))
+                CacheStatus::CacheSpecificError(String::from("download was cancelled"))
             );
             assert_eq!(server.accesses(), 0);
         })